requirements.txt
streamlit
anthropic
python-dotenv
openai-whisper
soundfile
numpy
torch
edge-tts
SpeechRecognition
pyaudio
emoji
pyautogui
PyQt6
opencv-python
mediapipe
